+++ packages/editor/editor-core/src/__tests__/integration/status/emoji-with-mention.ts
@@ -14,7 +14,7 @@ import {
 BrowserTestCase(
   'emoji.ts: Insert an emoji, then a mention, move to right before the emoji and try to add text between both',
   { skip: ['ie'] },
-  async (client: any) => {
+  async (client: any, testName: string) => {
     const browser = new Page(client);
 
     await browser.goto(fullpage.path);
@@ -46,6 +46,6 @@ BrowserTestCase(
 
     await browser.click(editable);
     const doc = await browser.$eval(editable, getDocFromElement);
-    expect(doc).toMatchDocSnapshot();
+    expect(doc).toMatchCustomDocSnapshot(testName);
   },
 );