+++ packages/editor/editor-core/src/__tests__/integration/status/panel-with-status.ts
@@ -11,7 +11,7 @@ import {
 BrowserTestCase(
   'status.ts: Insert status into panel, move cursor to right before status, and add text',
   { skip: ['ie'] },
-  async (client: any) => {
+  async (client: any, testName: string) => {
     const browser = new Page(client);
 
     await browser.goto(fullpage.path);
@@ -33,14 +33,14 @@ BrowserTestCase(
     ]);
 
     const doc = await browser.$eval(editable, getDocFromElement);
-    expect(doc).toMatchDocSnapshot();
+    expect(doc).toMatchCustomDocSnapshot(testName);
   },
 );
 
 BrowserTestCase(
   'status.ts: Insert status into panel, move cursor to right before panel, move right, and add text',
   { skip: ['ie'] },
-  async (client: any) => {
+  async (client: any, testName: string) => {
     const browser = new Page(client);
 
     await browser.goto(fullpage.path);
@@ -64,6 +64,6 @@ BrowserTestCase(
     ]);
 
     const doc = await browser.$eval(editable, getDocFromElement);
-    expect(doc).toMatchDocSnapshot();
+    expect(doc).toMatchCustomDocSnapshot(testName);
   },
 );