+++ packages/editor/editor-core/src/__tests__/visual-regression/table/block-alignment.ts
@@ -9,6 +9,7 @@ import {
   setTableLayout,
   getSelectorForTableCell,
 } from '../../__helpers/page-objects/_table';
+import { waitForEmojis } from '../../__helpers/page-objects/_emoji';
 
 describe('Table with block looks correct for fullpage:', () => {
   let page: any;
@@ -19,6 +20,7 @@ describe('Table with block looks correct for fullpage:', () => {
   });
 
   afterEach(async () => {
+    await waitForEmojis(page);
     await snapshot(page, 0.01);
   });
 
@@ -31,12 +33,14 @@ describe('Table with block looks correct for fullpage:', () => {
     await initFullPageEditorWithAdf(page, adf, Device.LaptopMDPI);
     await page.click(getSelectorForTableCell({ row: 4, cell: 1 }));
     await setTableLayout(page, 'wide');
+    await page.click(getSelectorForTableCell({ row: 4, cell: 1 }));
   });
 
   it('full-width layout ', async () => {
     await initFullPageEditorWithAdf(page, adf, Device.LaptopHiDPI);
     await page.click(getSelectorForTableCell({ row: 4, cell: 1 }));
     await setTableLayout(page, 'full-width');
+    await page.click(getSelectorForTableCell({ row: 4, cell: 1 }));
   });
 });