+++ packages/editor/editor-core/src/plugins/media/ui/ResizableMediaSingle/styled.tsx
@@ -10,7 +10,6 @@ export const Wrapper: React.ComponentClass<
   & > div {
     ${MediaSingleDimensionHelper};
     position: relative;
-    z-index: 1;
 
     > div {
       position: absolute;