+++ packages/editor/editor-mobile-bridge/src/__tests__/integration/native-to-web/links.ts
@@ -14,7 +14,7 @@ import {
 BrowserTestCase(
   `links.ts: Insert link on empty content`,
   { skip },
-  async client => {
+  async (client: any, testName: string) => {
     const browser = new Page(client);
 
     await browser.goto(editor.path);
@@ -28,7 +28,7 @@ BrowserTestCase(
     );
 
     const doc = await browser.$eval(editable, getDocFromElement);
-    expect(doc).toMatchDocSnapshot();
+    expect(doc).toMatchCustomDocSnapshot(testName);
   },
 );
 
@@ -36,7 +36,7 @@ BrowserTestCase(
   `links.ts: Insert link on existing text node`,
   // Safari has issues with key events
   { skip: skip.concat('safari') },
-  async client => {
+  async (client: any, testName: string) => {
     const browser = new Page(client);
 
     await browser.goto(editor.path);
@@ -53,7 +53,7 @@ BrowserTestCase(
     );
 
     const doc = await browser.$eval(editable, getDocFromElement);
-    expect(doc).toMatchDocSnapshot();
+    expect(doc).toMatchCustomDocSnapshot(testName);
   },
 );
 
@@ -61,7 +61,7 @@ BrowserTestCase(
   `links.ts: Insert link with text selection`,
   // Safari has issues with key events
   { skip: skip.concat('safari') },
-  async client => {
+  async (client: any, testName: string) => {
     const browser = new Page(client);
 
     await browser.goto(editor.path);
@@ -81,7 +81,7 @@ BrowserTestCase(
     );
 
     const doc = await browser.$eval(editable, getDocFromElement);
-    expect(doc).toMatchDocSnapshot();
+    expect(doc).toMatchCustomDocSnapshot(testName);
   },
 );
 
@@ -89,7 +89,7 @@ BrowserTestCase(
   `links.ts: change link text`,
   // Safari has issues with key events
   { skip: skip.concat('safari') },
-  async client => {
+  async (client: any, testName: string) => {
     const browser = new Page(client);
 
     await browser.goto(editor.path);
@@ -112,7 +112,7 @@ BrowserTestCase(
     );
 
     const doc = await browser.$eval(editable, getDocFromElement);
-    expect(doc).toMatchDocSnapshot();
+    expect(doc).toMatchCustomDocSnapshot(testName);
   },
 );
 
@@ -120,7 +120,7 @@ BrowserTestCase(
   `links.ts: change link href`,
   // Safari has issues with key events
   { skip: skip.concat('safari') },
-  async client => {
+  async (client: any, testName: string) => {
     const browser = new Page(client);
 
     await browser.goto(editor.path);
@@ -143,7 +143,7 @@ BrowserTestCase(
     );
 
     const doc = await browser.$eval(editable, getDocFromElement);
-    expect(doc).toMatchDocSnapshot();
+    expect(doc).toMatchCustomDocSnapshot(testName);
   },
 );
 
@@ -151,7 +151,7 @@ BrowserTestCase(
   `links.ts: Remove link`,
   // Safari has issues with key events
   { skip: skip.concat('safari') },
-  async client => {
+  async (client: any, testName: string) => {
     const browser = new Page(client);
 
     await browser.goto(editor.path);
@@ -178,6 +178,6 @@ BrowserTestCase(
     await callNativeBridge(browser, 'onLinkUpdate', 'text', '');
 
     const doc = await browser.$eval(editable, getDocFromElement);
-    expect(doc).toMatchDocSnapshot();
+    expect(doc).toMatchCustomDocSnapshot(testName);
   },
 );