+++ packages/elements/user-picker/src/components/Option.tsx
@@ -1,6 +1,6 @@
 import { components } from '@atlaskit/select';
 import * as React from 'react';
-import { Option as OptionType, UserPickerProps } from '../types';
+import { Option as OptionType } from '../types';
 import { EmailOption } from './EmailOption';
 import { TeamOption } from './TeamOption';
 import { UserOption } from './UserOption';
@@ -10,7 +10,9 @@ export type OptionProps = {
   data: OptionType;
   isSelected: boolean;
   status?: string;
-  selectProps: UserPickerProps;
+  selectProps: {
+    emailLabel?: string;
+  };
 };
 
 const dataOption = ({