+++ packages/search/global-search/src/components/common/QuickSearchContainer.tsx
@@ -26,6 +26,12 @@ import { ABTest } from '../../api/CrossProductSearchClient';
 const resultMapToArray = (results: GenericResultMap): Result[][] =>
   objectValues(results).reduce((acc: Result[][], value) => [...acc, value], []);
 
+const DEFAULT_AB_TEST: ABTest = Object.freeze({
+  experimentId: 'default',
+  abTestId: 'default',
+  controlId: 'default',
+});
+
 export interface SearchResultProps extends State {
   retrySearch: Function;
 }
@@ -58,6 +64,7 @@ export interface Props {
   placeholder?: string;
   selectedResultId?: string;
   onSelectedResultIdChanged?: (id: string) => void;
+  enablePreQueryFromAggregator?: boolean;
 }
 
 export interface State {
@@ -68,6 +75,7 @@ export interface State {
   keepPreQueryState: boolean;
   searchResults: GenericResultMap | null;
   recentItems: GenericResultMap | null;
+  abTest: ABTest;
 }
 
 const LOGGER_NAME = 'AK.GlobalSearch.QuickSearchContainer';
@@ -92,6 +100,7 @@ export class QuickSearchContainer extends React.Component<Props, State> {
       recentItems: null,
       searchResults: null,
       keepPreQueryState: true,
+      abTest: DEFAULT_AB_TEST,
     };
   }
 
@@ -154,6 +163,7 @@ export class QuickSearchContainer extends React.Component<Props, State> {
               timings || {},
               this.state.searchSessionId,
               this.state.latestSearchQuery,
+              this.state.abTest,
             );
           },
         );
@@ -172,11 +182,22 @@ export class QuickSearchContainer extends React.Component<Props, State> {
     }
   };
 
-  fetchAbTestData = async (searchSessionId: string) => {
+  fetchAbTestDataOrDefault = async (searchSessionId: string) => {
     const { getAbTestData } = this.props;
     const startTime = performanceNow();
-    const abTest = await getAbTestData(searchSessionId);
+
+    let abTest;
+    try {
+      abTest = (await getAbTestData(searchSessionId)) || DEFAULT_AB_TEST;
+    } catch (error) {
+      abTest = DEFAULT_AB_TEST;
+    }
+
     const elapsedMs = performanceNow() - startTime;
+
+    this.setState({
+      abTest,
+    });
 
     return {
       elapsedMs,
@@ -210,11 +231,16 @@ export class QuickSearchContainer extends React.Component<Props, State> {
   fireShownPreQueryEvent = (
     searchSessionId,
     recentItems,
+    abTest: ABTest,
     requestStartTime?: number,
     experimentRequestDurationMs?: number,
     renderStartTime?: number,
   ) => {
-    const { createAnalyticsEvent, getDisplayedResults } = this.props;
+    const {
+      createAnalyticsEvent,
+      getDisplayedResults,
+      enablePreQueryFromAggregator,
+    } = this.props;
     if (createAnalyticsEvent && getDisplayedResults) {
       const elapsedMs: number = requestStartTime
         ? performanceNow() - requestStartTime
@@ -227,9 +253,9 @@ export class QuickSearchContainer extends React.Component<Props, State> {
       const resultsArray: Result[][] = resultMapToArray(
         getDisplayedResults(recentItems),
       );
-      const eventAttributes: ShownAnalyticsAttributes = buildShownEventDetails(
-        ...resultsArray,
-      );
+      const eventAttributes: ShownAnalyticsAttributes = {
+        ...buildShownEventDetails(...resultsArray),
+      };
 
       firePreQueryShownEvent(
         eventAttributes,
@@ -237,7 +263,9 @@ export class QuickSearchContainer extends React.Component<Props, State> {
         renderTime,
         searchSessionId,
         createAnalyticsEvent,
+        abTest,
         experimentRequestDurationMs,
+        !!enablePreQueryFromAggregator,
       );
     }
   };
@@ -249,6 +277,7 @@ export class QuickSearchContainer extends React.Component<Props, State> {
     timings,
     searchSessionId,
     latestSearchQuery: string,
+    abTest: ABTest,
   ) => {
     const performanceTiming: PerformanceTiming = {
       startTime,
@@ -264,13 +293,13 @@ export class QuickSearchContainer extends React.Component<Props, State> {
       const resultsDetails: ShownAnalyticsAttributes = buildShownEventDetails(
         ...resultsArray,
       );
-
       firePostQueryShownEvent(
         resultsDetails,
         performanceTiming,
         searchSessionId,
         latestSearchQuery,
         createAnalyticsEvent,
+        abTest,
       );
     }
   };
@@ -284,6 +313,8 @@ export class QuickSearchContainer extends React.Component<Props, State> {
     }
 
     if (newLatestSearchQuery.length === 0) {
+      const { abTest } = this.state;
+
       // reset search results so that internal state between query and results stays consistent
       this.setState(
         {
@@ -295,6 +326,7 @@ export class QuickSearchContainer extends React.Component<Props, State> {
           this.fireShownPreQueryEvent(
             this.state.searchSessionId,
             this.state.recentItems || {},
+            abTest,
           ),
       );
     } else {
@@ -315,7 +347,9 @@ export class QuickSearchContainer extends React.Component<Props, State> {
       });
     }
 
-    const abTestPromise = this.fetchAbTestData(this.state.searchSessionId);
+    const abTestPromise = this.fetchAbTestDataOrDefault(
+      this.state.searchSessionId,
+    );
     this.fireExperimentExposureEvent(
       this.state.searchSessionId,
       abTestPromise.then(({ abTest }) => abTest),
@@ -335,10 +369,15 @@ export class QuickSearchContainer extends React.Component<Props, State> {
           isLoading: false,
         },
         async () => {
-          const experimentRequestDurationMs = (await abTestPromise).elapsedMs;
+          const {
+            elapsedMs: experimentRequestDurationMs,
+            abTest,
+          } = await abTestPromise;
+
           this.fireShownPreQueryEvent(
             this.state.searchSessionId,
             this.state.recentItems || {},
+            abTest,
             startTime,
             experimentRequestDurationMs,
             renderStartTime,
@@ -383,6 +422,7 @@ export class QuickSearchContainer extends React.Component<Props, State> {
       searchResults,
       recentItems,
       keepPreQueryState,
+      abTest,
     } = this.state;
 
     return (
@@ -407,6 +447,7 @@ export class QuickSearchContainer extends React.Component<Props, State> {
           recentItems,
           keepPreQueryState,
           searchSessionId,
+          abTest,
         })}
       </GlobalQuickSearch>
     );